aligned integer types for BRGEMM frontend

+++ src/libxs_main.c
@@ -3008,8 +3008,8 @@ LIBXS_API libxs_gemmfunction libxs_dispatch_brgemm_v2( const libxs_gemm_shape ge
     desc->c3 = (unsigned char)(((brgemm_config.br_unroll_hint < 255) && (brgemm_config.br_unroll_hint > 0)) ? brgemm_config.br_unroll_hint : 0);
   }
   if ( brgemm_config.br_type == LIBXS_GEMM_BATCH_REDUCE_STRIDE ) {
-    desc->c1 = brgemm_config.br_stride_a_hint;
-    desc->c2 = brgemm_config.br_stride_b_hint;
+    desc->c1 = (long long)brgemm_config.br_stride_a_hint;
+    desc->c2 = (long long)brgemm_config.br_stride_b_hint;
   }
 
   /* JIT! */
@@ -3058,8 +3058,8 @@ LIBXS_API libxs_gemmfunction_ext libxs_dispatch_brgemm_ext_v2( const libxs_gemm_
     desc->c3 = (unsigned char)(((brgemm_config.br_unroll_hint < 255) && (brgemm_config.br_unroll_hint > 0)) ? brgemm_config.br_unroll_hint : 0);
   }
   if ( brgemm_config.br_type == LIBXS_GEMM_BATCH_REDUCE_STRIDE ) {
-    desc->c1 = brgemm_config.br_stride_a_hint;
-    desc->c2 = brgemm_config.br_stride_b_hint;
+    desc->c1 = (long long)brgemm_config.br_stride_a_hint;
+    desc->c2 = (long long)brgemm_config.br_stride_b_hint;
   }
 
   /* setting binary post-op eltwise fields */
@@ -4028,8 +4028,8 @@ LIBXS_API libxs_dmmfunction_reducebatch_strd libxs_dmmdispatch_reducebatch_strd(
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4051,8 +4051,8 @@ LIBXS_API libxs_smmfunction_reducebatch_strd libxs_smmdispatch_reducebatch_strd(
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4074,8 +4074,8 @@ LIBXS_API libxs_bsmmfunction_reducebatch_strd libxs_bsmmdispatch_reducebatch_str
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4097,8 +4097,8 @@ LIBXS_API libxs_bmmfunction_reducebatch_strd libxs_bmmdispatch_reducebatch_strd(
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4120,8 +4120,8 @@ LIBXS_API libxs_wimmfunction_reducebatch_strd libxs_wimmdispatch_reducebatch_str
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4143,8 +4143,8 @@ LIBXS_API libxs_ssbimmfunction_reducebatch_strd libxs_ssbimmdispatch_reducebatch
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4166,8 +4166,8 @@ LIBXS_API libxs_usbimmfunction_reducebatch_strd libxs_usbimmdispatch_reducebatch
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_A_UNSIGNED | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4189,8 +4189,8 @@ LIBXS_API libxs_subimmfunction_reducebatch_strd libxs_subimmdispatch_reducebatch
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_B_UNSIGNED | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4212,8 +4212,8 @@ LIBXS_API libxs_uubimmfunction_reducebatch_strd libxs_uubimmdispatch_reducebatch
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_AB_UNSIGNED | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4236,8 +4236,8 @@ LIBXS_API libxs_sububmmfunction_reducebatch_strd libxs_sububmmdispatch_reducebat
     gemm_flags | LIBXS_GEMM_FLAG_B_UNSIGNED | LIBXS_GEMM_FLAG_C_UNSIGNED | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE,
     libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4259,8 +4259,8 @@ LIBXS_API libxs_dmmfunction_reducebatch_strd libxs_dmmdispatch_reducebatch_strd_
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
@@ -4283,8 +4283,8 @@ LIBXS_API libxs_smmfunction_reducebatch_strd libxs_smmdispatch_reducebatch_strd_
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
@@ -4307,8 +4307,8 @@ LIBXS_API libxs_bsmmfunction_reducebatch_strd libxs_bsmmdispatch_reducebatch_str
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
@@ -4331,8 +4331,8 @@ LIBXS_API libxs_bmmfunction_reducebatch_strd libxs_bmmdispatch_reducebatch_strd_
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
@@ -4355,8 +4355,8 @@ LIBXS_API libxs_wimmfunction_reducebatch_strd libxs_wimmdispatch_reducebatch_str
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
@@ -4379,8 +4379,8 @@ LIBXS_API libxs_ssbimmfunction_reducebatch_strd libxs_ssbimmdispatch_reducebatch
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
@@ -4403,8 +4403,8 @@ LIBXS_API libxs_usbimmfunction_reducebatch_strd libxs_usbimmdispatch_reducebatch
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_A_UNSIGNED | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
@@ -4427,8 +4427,8 @@ LIBXS_API libxs_subimmfunction_reducebatch_strd libxs_subimmdispatch_reducebatch
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_B_UNSIGNED | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
@@ -4451,8 +4451,8 @@ LIBXS_API libxs_uubimmfunction_reducebatch_strd libxs_uubimmdispatch_reducebatch
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_AB_UNSIGNED | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
@@ -4476,8 +4476,8 @@ LIBXS_API libxs_sububmmfunction_reducebatch_strd libxs_sububmmdispatch_reducebat
     gemm_flags | LIBXS_GEMM_FLAG_B_UNSIGNED | LIBXS_GEMM_FLAG_C_UNSIGNED | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE,
     libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
@@ -4501,8 +4501,8 @@ LIBXS_API libxs_bmmfunction_reducebatch_strd_meltwfused libxs_bmmdispatch_reduce
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4531,8 +4531,8 @@ LIBXS_API libxs_bmmfunction_reducebatch_strd_meltwfused libxs_bmmdispatch_reduce
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
@@ -4562,8 +4562,8 @@ LIBXS_API libxs_bsmmfunction_reducebatch_strd_meltwfused libxs_bsmmdispatch_redu
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
   }
@@ -4592,8 +4592,8 @@ LIBXS_API libxs_bsmmfunction_reducebatch_strd_meltwfused libxs_bsmmdispatch_redu
     NULL != ldc ? *ldc : m, NULL != alpha ? *alpha : LIBXS_ALPHA, NULL != beta ? *beta : LIBXS_BETA,
     gemm_flags | LIBXS_GEMM_FLAG_BATCH_REDUCE_STRIDE, libxs_get_gemm_xprefetch(prefetch));
   /*const*/ libxs_xmmfunction result;
-  desc->c1 = (unsigned long long)stride_a;
-  desc->c2 = (unsigned long long)stride_b;
+  desc->c1 = (long long)stride_a;
+  desc->c2 = (long long)stride_b;
   desc->c3 = (unsigned char)(((unroll_hint < 255) && (unroll_hint > 0)) ? unroll_hint : 0);
   if ( (stride_a < 0) || (stride_b < 0) ) {
     return NULL;
